+++ kite-core/src/main/kotlin/com/tang/kite/handler/result/DefaultResultHandler.kt
@@ -5,7 +5,7 @@ import java.lang.reflect.Field
 /**
  * @author Tang
  */
-class DefaultResultHandler : ResultHandler {
+object DefaultResultHandler : ResultHandler {
 
     override fun <T> setValue(field: Field, instance: T, value: Any) {
         field.set(instance, value)