+++ kite-core/src/main/kotlin/com/tang/kite/handler/result/primitive/BooleanResultHandler.kt
@@ -8,7 +8,7 @@ import java.lang.reflect.Field
 /**
  * @author Tang
  */
-class BooleanResultHandler : ResultHandler {
+object BooleanResultHandler : ResultHandler {
 
     override fun <T> setValue(field: Field, instance: T, value: Any) {
         val boolean = when (value) {