+++ kite-core/src/main/kotlin/com/tang/kite/handler/result/primitive/ShortResultHandler.kt
@@ -8,7 +8,7 @@ import java.lang.reflect.Field
 /**
  * @author Tang
  */
-class ShortResultHandler : ResultHandler {
+object ShortResultHandler : ResultHandler {
 
     override fun <T> setValue(field: Field, instance: T, value: Any) {
         val short = when (value) {